fix(stylex_utils): widen the structural key to 128 bits, on a faster hasher
Two of the five consumers of the evaluator's structural key act on a hash hit
with nothing confirming it. `evaluate_cached_within_budget` returns a cached fold
straight off `seen.get(&hash)`, and `flush_pending_insertions` splices a
declaration's style metadata off `before_decl.remove(&hash)`. For those two the
key is not a bucket, it is the equality test -- so at 64 bits and ten thousand
distinct expressions in a file, roughly one file in `1e12` gets a wrong folded
value or metadata spliced before the wrong declaration. Silently, in the output,
with no diagnostic. The `stylex-utils` glossary said all of them confirmed.

128 bits puts that past `1e-31`. Getting there with the standard library means
two salted `DefaultHasher` states fed by one walk, which was built and measured
first and cost +49% on the key and +5.8% on a whole production transform of a
400-`create` file. Paying that forever against a failure that arrives once per
`1e4` years is the wrong trade. xxh3 emits 128 bits from a single pass instead of
hashing the same bytes twice, and is enough faster than SipHash that the wider
key is also the cheaper one -- one key 30-43% faster, a deep fold 29% faster, the
whole transform of that file unchanged at 25.9 ms against 26.0.

Every consumer's key widened with it: `seen`, `jsx_spread_attr_exprs_map`,
`queued_decl_items`, `all_call_expressions` and `InsertionSlot::BeforeDecl`. The
last of those was the fifth consumer, found while counting -- its reads always
confirmed with `eq_ignore_span`, but `replace_call_expression` removes by hash
alone and would have evicted the wrong entry.

No output changes, and none could: nothing persists a key, nothing derives a
class name from one, and injection order comes from source order rather than from
hash order. `DefaultHasher` was never stable across Rust releases either, so
nothing could have depended on the old values and been correct. The fallback
cases now compare against `stable_hash_wide`, so an arm that quietly narrowed
back to 64 bits fails to compile.

Adds one direct dependency, `xxhash-rust` with the `xxh3` feature, BSD-2-Clause
and already on the licence allow-list.

+++ crates/stylex-transform/src/shared/structures/state_manager.rs
@@ -116,7 +116,7 @@ pub(crate) enum InsertionSlot {
   /// by the stable hash of the originating var-decl initializer.
   /// Emitted immediately before the matching declarator. Replaces the
   /// legacy `styles_to_inject` map.
-  BeforeDecl(u64),
+  BeforeDecl(u128),
 }
 
 #[derive(Debug, Clone)]
@@ -245,7 +245,7 @@ impl ModuleSourceState {
 
 #[derive(Clone, Debug, Default)]
 pub(crate) struct CallExpressionState {
-  all_call_expressions: FxHashMap<u64, Callee>,
+  all_call_expressions: FxHashMap<u128, Callee>,
 }
 
 impl CallExpressionState {
@@ -353,7 +353,7 @@ pub(crate) struct StyleInjectionState {
   /// per-bucket `Vec` small (typically 1–2 entries), so
   /// `Vec::contains` short-circuits early on PartialEq mismatches.
   /// Replaces the legacy `styles_to_inject` field's dual role.
-  queued_decl_items: IndexMap<u64, Vec<ModuleItem>>,
+  queued_decl_items: IndexMap<u128, Vec<ModuleItem>>,
 }
 
 impl StyleInjectionState {
@@ -437,7 +437,7 @@ pub struct StateManager {
   /// hoists its result out of a nested position must not hoist here.
   pub(crate) pattern_bound_top_level_calls: FxHashSet<Span>,
   pub(crate) call_expressions: CallExpressionState,
-  pub(crate) seen: FxHashMap<u64, Rc<SeenValue>>,
+  pub(crate) seen: FxHashMap<u128, Rc<SeenValue>>,
   /// How many expression levels the evaluator is currently inside.
   ///
   /// Lives here rather than on `EvaluationState` because the evaluator's
@@ -458,7 +458,7 @@ pub struct StateManager {
   /// attributes to a structurally-different spread — the hash only narrows the
   /// candidate set; equality decides. Buckets are size-1 in the absence of a
   /// collision, so the extra check is a single comparison.
-  pub(crate) jsx_spread_attr_exprs_map: FxHashMap<u64, Vec<(Expr, Vec<JSXAttrOrSpread>)>>,
+  pub(crate) jsx_spread_attr_exprs_map: FxHashMap<u128, Vec<(Expr, Vec<JSXAttrOrSpread>)>>,
 
   // `stylex.create` calls
   pub(crate) style_map: FxHashMap<String, Rc<StylesObjectMap>>,
@@ -1791,7 +1791,7 @@ pub(crate) fn flush_pending_insertions(
   let mut before_imports: Vec<ModuleItem> = Vec::new();
   let mut theme_imports: Vec<ModuleItem> = Vec::new();
   let mut after_imports: Vec<ModuleItem> = Vec::new();
-  let mut before_decl: FxHashMap<u64, Vec<ModuleItem>> = FxHashMap::default();
+  let mut before_decl: FxHashMap<u128, Vec<ModuleItem>> = FxHashMap::default();
 
   for PendingInsertion { slot, item } in pending {
     match slot {
@@ -1894,8 +1894,8 @@ pub(crate) fn flush_pending_insertions(
 /// Stable hashes of every relevant var-decl initializer reachable from
 /// `item`, matching the keys [`StateManager::queue_insertion`] uses
 /// under [`InsertionSlot::BeforeDecl`].
-fn decl_init_hashes(item: &ModuleItem) -> Vec<u64> {
-  let mut hashes: Vec<u64> = Vec::new();
+fn decl_init_hashes(item: &ModuleItem) -> Vec<u128> {
+  let mut hashes: Vec<u128> = Vec::new();
 
   let var_decls: Option<Vec<&VarDeclarator>> = match item {
     ModuleItem::ModuleDecl(ModuleDecl::ExportDecl(export_decl)) => export_decl
